fix(clients): file name changed and the provider edits
